IoT-Cardiac-Arrest-Self-Driving/requirements.txt
azure-iot-device==2.9.0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
decorator==5.1.0
deprecation==2.1.0
future==0.18.2
geocoder==1.38.1
idna==3.3
janus==1.0.0
packaging==21.3
paho-mqtt==1.6.1
pyparsing==3.0.6
PySocks==1.7.1
ratelim==0.1.6
requests==2.26.0
requests-unixsocket==0.3.0
six==1.16.0
typing-extensions==4.0.1
urllib3==1.26.7
